==> 1_Data preprocessing.ipynb <==
#Import the libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import scipy as sci

import sklearn.ensemble as se
import sklearn.model_selection as ms
import sklearn.metrics as sm
import sklearn.preprocessing as sp
import sklearn.impute as si
import seaborn as sns

import math

#load the raw dataset
pd.set_option('display.max_columns',None)
rawdata = pd.read_excel('Rawdata.xlsx')
rawdata.describe(include='all')

rawdata.columns

#Missing value of each feature and target
rawdata.iloc[:,2:27].isnull().sum()

#Boxplot 
a=0
fig = plt.figure(figsize=(18,20))
#data1 = pd.concat([rawdata.iloc[:,7:10],rawdata.iloc[:,11:18]],axis=1)
data1 = pd.concat([rawdata.iloc[:,2:26]],axis=1)
for i in range(24):
    axes = fig.add_subplot(4,6,a+1)
    axes = sns.boxplot(y=data1.iloc[:,i],orient="v",color="darkgreen",width=0.5)
    axes.tick_params(labelsize=12)
    axes.set_ylabel(data1.columns[i],size=12)
    plt.subplots_adjust(wspace=0.4, hspace=0.1)
    a=a+1
plt.savefig("figure_1.pdf")

#Correlation heatmap
plt.figure(figsize=(10,8))
h = sns.heatmap(data1.corr(method="spearman"), vmax=1, vmin=-0.8,
            cmap='YlGn',cbar=True,
            annot=True, fmt='.2f', annot_kws={'size':8, 'fontname':'Arial'})
plt.xticks(rotation = 90, fontsize= 10, fontname = "Arial")
plt.yticks(rotation = 0, fontsize= 10, fontname = "Arial")
#plt.show()
plt.savefig("figure_2.pdf")

#Extract the feature for the following missing value imputation
df = rawdata.iloc[:, 4:17]
df.describe()

## Missforest imputation

len(df.columns)

train_idx = []
test_idx = []
df_iter = []
gamma = []
impute_df = []
for b in range(len(df.columns)):
    #Extract the index of the missing value and non-missing value
    train_idx.append(df.iloc[:,b].index[np.where(df.iloc[:,b].notnull())[0]])
    test_idx.append(df.iloc[:,b].index[np.where(df.iloc[:,b].isnull())[0]])

sortindex = np.argsort(df.isnull().sum(axis=0)).values
df_miss = df
#Initialize the stopping criterion
gp=[1,1,1,1,1,1,1,1,1,1,1,1,1]
#Set the iterations 
for a in range(5):
    print("iterations:{}".format(a))
    #In the first imputation, the dataset is imputed by the mean values
    if a < 1:
        df_z = si.SimpleImputer(missing_values = np.nan, strategy='mean').fit_transform(df_miss)
        df_z = pd.DataFrame(df_z, columns = df_miss.columns)
    #After the first imputation, the dataset is replaced by the former imputed dataset  
    else:
        df_z = df_miss.copy()
    
    for i in np.delete(sortindex,0):
        fillc = df_z.iloc[:,i]
        df_new = df_z.iloc[:,df_z.columns != df_z.columns[i]]     
        #Split the dataset
        Ytrain = fillc[train_idx[i]]
        Ytest = fillc[test_idx[i]]
        Xtrain = df_new.loc[train_idx[i],:]
        Xtest = df_new.loc[test_idx[i],:]  
        #Apply the random forest with defualt hyperparameters to impute the missing values
        rfr_impute = se.RandomForestRegressor(random_state=42,oob_score=True)
        rfr_impute = rfr_impute.fit(Xtrain, Ytrain)  
        #Apply cross-validation to explore the imputation performance 
        cv = ms.KFold(n_splits=5, shuffle=True, random_state=42)
        scores = ms.cross_val_score(rfr_impute, Xtrain, Ytrain , cv=cv, scoring='r2')
        cv_R2 = np.mean(scores)
        #Impute
        Ypredict = rfr_impute.predict(Xtest)
        impute_df.append(Ypredict)
        #The imputed feature is filled in the raw dataset
        df_z.loc[test_idx[i],df_z.columns[i]]=Ypredict 
        print(df_miss.columns[i])
        print("Train R2: %.4f" %(rfr_impute.score(Xtrain,Ytrain)))  
        print("CV R2: %.4f (+/- %.4f)" %(cv_R2, scores.std())) 
        print("Oob score: %.4f" %(rfr_impute.oob_score_)) 
        
        
    #Examine the change of the stopping criterion after imputation
    if a < 1:
        yita = sum(np.array((df_z - df_miss.fillna(value=0))**2))/ sum((np.array(df_z)**2))
    else:
        yita = sum(np.array((df_z - df_miss)**2))/ sum((np.array(df_z)**2))
    print(yita)

    if ((yita - gp) >  0).any():
        break
    gp = yita 
    gamma.append(yita)
    df_iter.append(df_z)
    df_miss = df_z
print("The iteration is converged in:{}".format(a-1))

df_iter

#Plot the change of data distribution after each imputation
a=0
implabel=[ 'Depth', 'Temperature', 'Salinity', 'Density', 'O2', 'AT', 'CT', 'pH', 'pCO2', 'NO3', 'PO4', 'SiOH4', 'N2O']
fig = plt.figure(figsize=(15,8))
for i in sortindex:
    axes = fig.add_subplot(3,5,a+1)
    axes = sns.kdeplot(df.iloc[:,i],color='black',linewidth=1, linestyle='--')
    axes = sns.kdeplot(df_iter[0].iloc[:,i],color='green',linewidth=1)
    axes = sns.kdeplot(df_iter[1].iloc[:,i],color='darkgreen',linewidth=2)
    axes.tick_params(labelsize=14)
    axes.set_xlabel(implabel[i],size=16) 
    axes.set_ylabel('Density',size=16)
    plt.subplots_adjust(wspace=0.32, hspace=0.3)
    a=a+1
plt.legend(['raw data','1st imputed data', '2nd imputed data'],fontsize=10)
plt.savefig("figure_3.pdf")

#Based on the iterative imputation, the data with second imputation is extracted 
df_iter[1].describe()

#Rearrange the dataset
df1 = pd.concat([rawdata.iloc[:,:4],df_iter[1],rawdata.iloc[:,17:26]],axis=1)
#Delete the bad data point
df2 = df1
display(df2.columns)
display(df2.describe(include='all'))

#Export the preprocessed data
df2.to_excel('Preprocessed_data.xlsx')

#Missing value of each feature and target
df2.iloc[:,2:27].isnull().sum()
