==> 3_Developed model.ipynb <==
#Import libraries (please check whether you have installed these libraries)
import numpy as np
import pandas as pd
import pickle

## Multiple Output Model for predicting the apparent quantum yields of PPRIs 
## photochemically generated by DOM

#Define the Multiple output model which can simultaneously predict the lnΦ3DOM* ,lnΦ1O2 and lnΦ·OH
class MultipleOutputModel():

    #Load the developed models     
    def __init__(self):
        with open('best_cbr_SP.pickle', 'rb') as e:
            self.model_tri_sp = pickle.load(e)
        with open('best_cbr_15N.pickle', 'rb') as f:
            self.model_tri_15n = pickle.load(f)
        with open('best_cbr_18O.pickle', 'rb') as g:
            self.model_tri_18o = pickle.load(g)
        
    
    #Make prediction
    def predict(self, x, regressorchain = False, export=False):
        newx = x.copy()
        x1 = newx.iloc[:,5:21]
        x2 = newx.iloc[:,5:21]
        x3 = newx.iloc[:,5:21]
        
        if regressorchain == True: 
            x2['Pred 15N'] = self.model_tri.predict(x1)
            x3['Pred 18O'] = self.model_tri.predict(x1)
            mult_x=[self.model_tri.predict(x1),self.model_tri_sin.predict(x2),self.model_tri_hyd.predict(x3)]
        else:
            mult_x=[self.model_tri_sp.predict(x1),self.model_tri_15n.predict(x2),self.model_tri_18o.predict(x3)]   
            
        df_x = pd.DataFrame(mult_x,index=['Pred SP' ,'Pred 15N', 'Pred 18O'])
        #export: default=False. Whether the predicted PPRIs is exported after the prediction.
        #If True, the predicted data will be exported into an Excel file.
        if export == True:
            df_x.T.to_excel('predicted N2O.xlsx')#you can design your path to export the Excel file.
            print('predicted N2O.xlsx is exported')
            display(df_x.T) 
            return mult_x
        else:
            display(df_x.T)
            return mult_x
    
    #Calculate R2 and RMSE for each PPRIs, if you have already calculated the observed PPRIs through photochemical experiments
    def mult_reg_score(self, true_y, pred_y, export = False):
        true_y1 = list(np.array(true_y.T))
        mult_r = []
        mult_rmse = []
        for i in range(len(pred_y)):
            y_mean = np.mean(true_y1[i])
            sse = sum((true_y1[i] - pred_y[i])**2)
            sst = sum((true_y1[i] - y_mean)**2) 
            r2 = 1 - (sse/sst)
            mult_r.append(r2)
            rmse = np.mean((true_y1[i] - pred_y[i])**2) **0.5
            mult_rmse.append(rmse)
        mult_result = pd.DataFrame({'R2':mult_r,'RMSE':mult_rmse},index=['SP' ,'15N', '18O'])
        #export: default=False. Whether the calculated R2 and RMSE is exported.
        #If True, he calculated R2 and RMSE will be exported into an Excel file.
        if export == True:
            mult_result = pd.to_excel('mult_result.xlsx')#you can design your path to export the Excel file.
            print('mult_result.xlsx is exported')
            return mult_result
        return mult_result

#Import the example data from a previous literature
#The data you prepared must sort by the feature order in the example data and 
#the feature should be converted into the uniform units and calculation
feature = pd.read_excel('out_surface_n2o.xlsx',sheet_name=0)#read the first sheet of file Example data.xlsx
feature.describe(include='all')#Statistical description of the example data

#Instantiating the Multiple Output Model
model = MultipleOutputModel()

predicted_y = model.predict(feature,regressorchain=False,export=True)

#If you have already calculated the PPRIs data and want to explore predictive performance of the developed models
#you can also import the observed PPRIs data to calculate the R2 and RMSE, but remember to transform PPRIs data into PPRIs data
target = pd.read_excel('out_surface_n2o.xlsx',sheet_name=0)
results = model.mult_reg_score(lntarget,predicted_y,export=False)
results
